requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1
django-crispy-forms==1.8.0
django-phone-field==1.8.0
get==2019.4.13
gunicorn==20.0.0
heroku==0.1.4
idna==2.8
Pillow==6.2.1
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
python-dateutil==1.5
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
urllib3==1.25.6

